+++ light-client/tests/model_based.rs
@@ -119,7 +119,8 @@ struct HeaderVersionFuzzer {}
 impl SingleStepTestFuzzer for HeaderVersionFuzzer {
     // TODO: rehash the header and re-compute commit with it
     // TODO: Unlike in tendermint-go, we don't assert for a particular version in rust
-    // TODO: Either add this check in verification or remove this test because otherwise there's no point of it
+    // TODO: Either add this check in verification or remove this test because otherwise there's no
+    // point of it
     fn fuzz_input(input: &mut BlockVerdict) -> (String, bool) {
         let mut rng = rand::thread_rng();