+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   deemed valid instead of invalid ([#650])
 - `[light-client]` Revert a change introduced in [#652] that would enable DoS attacks,
   where full nodes could spam the light client with massive commits (eg. 10k validators).
+- `[tendermint]` (Since RC1) CanonicalBlockId is now correctly decoded to `None` in `CanonicalVote` and `CanonicalProposal`.
 
 ### FEATURES:
 
@@ -25,6 +26,7 @@
 [#652]: https://github.com/informalsystems/tendermint-rs/pulls/652
 [#639]: https://github.com/informalsystems/tendermint-rs/pull/639
 [#660]: https://github.com/informalsystems/tendermint-rs/issues/660
+[#663]: https://github.com/informalsystems/tendermint-rs/issues/663
 
 ## v0.17.0-rc1